--- ACTIVIDAD_3/cancelar_pedido.php
@@ -1,5 +1,0 @@
-<?php 
-	unlink("nombre.csv");
-	unlink("pedido.csv");
-	header("Location:index.php");
- ?>
