--- ACTIVIDAD_3/header.php
@@ -1,8 +1,0 @@
-<nav class="nav">
-	<h2>Titulo</h2>
-	<ul class="nav-menu">
-		<li class="nav-menu-item"><a href="index.php">Inicio</a></li>
-		<li class="nav-menu-item"><a href="pedido.php">Pedido</a></li>
-		<li class="nav-menu-item"><a href="select.php">Estadisticas</a></li>
-	</ul>
-</nav>
