--- ACTIVIDAD_3/productos.json
@@ -1,33 +1,0 @@
-[   
-    {
-    	"nombre": "Hamburguesa Simple", 
-    	"precio": 1000,
-    	"imagen":"https://cache-backend-mcd.mcdonaldscupones.com/media/image/product$kqX8TYcp/200/200/original?country=ar",
-        "ingredientes": [
-            "cheddar",
-            "ketchup"
-        ]
-    },
-    {
-    	"nombre": "Hamburguesa Doble", 
-    	"precio": 1500, 
-    	"imagen":"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcT6CjbDPbJGV8hn9vwwIvGuzXm2OOhcQQba89CQuduzfg&s",
-        "ingredientes": [
-            "chedar",
-            "ketchup",
-            "carne extra"
-        ]
-    },
-    {
-    	"nombre": "CheeseBurger", 
-   		"precio": 2000, 
-    	"imagen":"https://s7d1.scene7.com/is/image/mcdonalds/mcdonalds-triple-cheeseburger-april-promo:nutrition-calculator-tile?wid=822&hei=822&dpr=off",
-        "ingredientes": [
-            "cheddar",
-            "ketchup",
-            "bacon"
-        ]	
-    }
-
-    
-]
