--- README.md
@@ -1,3 +1,0 @@
-Sucursal 
-primero que nada debemos contar con la aplicacion xampp o un servidor como puede ser 000webhost
-una vez que tengamos esto, debemos descargar los archivos de este repositorio, subir los que sean de tipo PHP-HTML-JS al servidor (en caso de que sea xampp, deben ir dentro de la carpeta htdocs), luego los archivos sql debemos importarlos para crear las bases de datos
