+++ Vistas/select.php
@@ -1,0 +1,32 @@
+<?php include 'header.php' ?>
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Seleccionar</title>
+    <link rel="stylesheet" href="../Vistas/styles/styles.css">
+</head>
+<body>
+	<section>
+        <h1 style="text-align:center;">SUCURSALES</h1>
+        <form action="prueba_grafico.php" method="post" class="menu">            
+            <ul class="links">
+                <li>
+                    <input id="input_head" name="sucursal" type="submit" value="Quilmes">
+                    
+                </li>
+                <li>
+                    <input id="input_head" name="sucursal" type="submit" value="Grand Bourg">
+                    
+                </li>
+                <li>
+                    <input id="input_head" name="sucursal" type="submit" value="Tortuguitas">
+                
+                </li>
+            </ul>
+        </form>
+    </section>
+    	
+</body>
+</html>
