--- hi.txt
@@ -1,1 +1,0 @@
-Hola mundo!
